requirements.txt
azure_storage==0.37.0
fitz==0.0.1.dev2
numpy==1.24.3
pydantic==2.9.2
pytest==8.3.3
scikit_learn==1.3.0
scipy==1.14.1
sentence_transformers==2.2.2
torch==2.4.1
